# apps/backend/src/configuration/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ServisBot API",
    "version": "1.0.0",
    "description": "API for managing bots, workers, and logs"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Get all bots",
        "description": "Retrieve a paginated list of all bots",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of bots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBots"
                }
              }
            }
          }
        }
      }
    },
    "/{bid}": {
      "get": {
        "summary": "Get a bot by ID",
        "description": "Retrieve details of a specific bot",
        "parameters": [
          {
            "name": "bid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bot ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Bot details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bot"
                }
              }
            }
          },
          "404": {
            "description": "Bot not found"
          }
        }
      }
    },
    "/{bid}/workers": {
      "get": {
        "summary": "Get workers for a bot",
        "description": "Retrieve a paginated list of workers belonging to a specific bot",
        "parameters": [
          {
            "name": "bid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bot ID"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of workers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedWorkers"
                }
              }
            }
          }
        }
      }
    },
    "/{bid}/logs": {
      "get": {
        "summary": "Get logs for a bot",
        "description": "Retrieve a paginated list of logs for a specific bot",
        "parameters": [
          {
            "name": "bid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bot ID"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLogs"
                }
              }
            }
          },
          "404": {
            "description": "Bot not found (if no logs found)"
          }
        }
      }
    },
    "/{bid}/workers/{wid}/logs": {
      "get": {
        "summary": "Get logs for a worker",
        "description": "Retrieve a paginated list of logs for a specific worker of a bot",
        "parameters": [
          {
            "name": "bid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bot ID"
          },
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Worker ID"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLogs"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Bot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED",
              "PAUSED"
            ]
          },
          "created": {
            "type": "integer"
          }
        }
      },
      "Worker": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bot": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created": {
            "type": "integer"
          }
        }
      },
      "Log": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bot": {
            "type": "string"
          },
          "worker": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "created": {
            "type": "string"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "PaginatedBots": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bot"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "PaginatedWorkers": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Worker"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "PaginatedLogs": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Log"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      }
    }
  }
}